requirements.txt
streamlit 
PyPDF2 
faiss-cpu 
openai 
python-dotenv 
numpy
langchain
langchain-community
tiktoken